Add Loading Spinner To Loading Bar Header
This adds a spinner to the left side of the loading bar header. This
will allow programs using this loading bar to indicate to users that the
program is not frozen even if the bar doesn't move for extended periods
of time due to the length of the task being tracked.