Support Drawing Loading Bar After Loading Complete
Previously if you drew the loading bar after loading was done, you would
end up with 2 headers because we did not account for the line separator.
The bar will now check if the previously drawn bar had a line sperator
at the end and if so, it will move up a second line to redraw. This will
not handle screen size changes that result in extra lines. A solution
for that will likely be added later